Fixed script loading for sample edit
