+++ src/main/webapp/pages/samples/sample_edit.html
@@ -178,7 +178,7 @@
 </div>
 
 <th:block layout:fragment="scripts">
-    <script th:src="@{/resources/js/pages/samples/edit.js}"></script>
+    <script th:src="@{/dist/js/sample-edit.bundle.js}"></script>
 </th:block>
 </body>
 </html>
